Added the code for twitter scraping
I am scraping twitter account with handle 'narendramodi'.[Which can be changed quite easily and accordingly]
Also we have to hardcode the date range for this part (which we have to handle later while making the website)

As of now the code can easily generate an excel file and other things (database) with tweet info for modi(or any account) for that specific day range.
The excel file in particular can be easily used to search an keyword.

For installation guide, refer to the original project.
Note:1. Apply for Twitter API keys and be careful about answering there questions.
2. make a twitter application at twitter developers and use the API keys in this code.
3.Allow access to the API calls in the form of read, write and direct msg.
4. In this course, ur account might get locked , just unlock it before running the metadata scrapping mechanism.